Remove the mock objects and connect the routes to the database.
When someone makes a post to /shorten, the api will check whether the url is already stored in the database.
If it is already stored, then it will retrieve it and send back, otherweise it will store the url and return a json with the shorter and original url.
Making a get request to the shorter url will make the api look for the original url in the database.
If the url is found, then the client will be redirected to the original url, otherweise it will send a json back with 400 status code saying url not found.